makes a ship of size 1x3
Successfully creates a ship 3 units long in a random location that can be horizontal or vertical.

Known issues with this change: check() does not work.  Tried multiple things, unable to get it to successfully check guess_row and guess_col against ship_row and ship_col as lists.  There will also need to be a new if statement to determine if the game was won as a hit no longer indicates a win.

The ship's starting location is inside the board, but it can currently extend outside of it.  Nesting an if statement inside of the for that sets the coordinates could check if it is outside the board and do something if so.